+++ src/pf_imgui/elements/plots/types/Line.cpp
@@ -7,6 +7,6 @@
 namespace pf::ui::ig::plot_type {
 Line::Line(const std::string &elementName, const std::string &caption) : LabeledPlotData(elementName, caption) {}
 
-void Line::renderImpl() { ImPlot::PlotLine(getLabel().c_str(), xData.data(), yData.data(), xData.size()); }
+void Line::renderImpl() { ImPlot::PlotLine(getLabel().c_str(), xData.data(), yData.data(), static_cast<int>(xData.size())); }
 
 }// namespace pf::ui::ig::plot_type
