+++ src/pf_imgui/interface/ElementContainer.cpp
@@ -27,7 +27,7 @@ void ElementContainer::insertChild(std::unique_ptr<Element> child, std::size_t i
 #ifndef _MSC_VER// TODO: MSVC internal error
   if (index > childrenInOrder.size()) { throw InvalidArgumentException("Index out of bounds: {}", index); }
 #endif
-  childrenInOrder.insert(childrenInOrder.begin() + index, *child);
+  childrenInOrder.insert(childrenInOrder.begin() + static_cast<long long>(index), *child);
   children[child->getName()] = std::move(child);
 }
